+++ Example/iOS Example/ViewController.swift
@@ -15,7 +15,7 @@ class ViewController: UIViewController {
         super.viewDidLoad()
         // Do any additional setup after loading the view, typically from a nib.
         
-        let client = APIClient(apiKey: "FAKE-API-KEY")
+        let client = DarkSkyClient(apiKey: "FAKE-API-KEY")
         client.units = .auto
         client.language = .english
         client.getForecast(latitude: 30, longitude: 30, excludeFields: [.alerts, .currently, .daily, .flags, .minutely]) { (result) in